gmail-to-sheets/requirements.txt
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
beautifulsoup4==4.12.3
